backend/requirements.txt
fastapi==0.115.13
uvicorn==0.34.3
pandas==2.2.3
requests==2.32.4
python-decouple==3.8
python-dotenv==1.0.0
pydantic==2.10.4 